DRF/cars/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
boto3==1.24.66
botocore==1.27.66
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
daphne==3.0.2
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.8.0
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.1
outcome==1.2.0
parsel==1.6.0
Pillow==9.2.0
Protego==0.2.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
PyJWT==2.4.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
s3transfer==0.6.0
Scrapy==2.6.2
selenium==4.4.3
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tldextract==3.3.1
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.12
w3lib==2.0.1
wsproto==1.2.0
zope.interface==5.4.0
celery==5.2.7
gunicorn==20.1.0
uvicorn==0.18.3